Made a ENTSOE library. Portugal switched to ENTSOE.
